add upload help statements, improve tests, add upload configs

+++ .github/scripts/run_mpx_pipeline_metadata.sh
@@ -1,11 +1,7 @@
 #!/bin/bash
 set -eo pipefail
 
-# Create a Cache Dir
-mkdir -p conda_cache_dir
-
-# Get partial human ref genome
-gunzip $PWD/.github/ci-data/partial_hg38_ref.fa.gz
+# Test is run second so no need to re-create envs or unzip files
 
 # Run Illumina Pipeline
 nextflow run ./main.nf \
@@ -13,6 +9,7 @@ nextflow run ./main.nf \
     --cache ./conda_cache_dir \
     --directory $PWD/.github/ci-data/ \
     --human_ref $PWD/.github/ci-data/partial_hg38_ref.fa
+    --metadata_csv $PWD/.github/ci-data/test_metadata.csv
 
 # Check that the output is as expected
 # 1. Num Reads
@@ -25,9 +22,14 @@ if [[ `awk -F, '$1 == "Sample1" {print $6}' overall_sample_quality.csv` != 0.845
     echo "Incorrect output: Genome Completeness"
     exit 1
 fi
+# 3. Check metadata put in correctly
+if [[ `awk -F, '$1 == "Sample1" {print $7}' overall_sample_quality.csv` != 231 ]]; then 
+    echo "Incorrect output: Metadata not parsed correctly"
+    exit 1
+fi
 
 # Reset and Track
-mv .nextflow.log artifacts/nextflow.log
+mv .nextflow.log artifacts/nextflow2.log
 rm -rf results work/ .nextflow*
 
 echo "Passed Tests"
